pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trsllava"
version = "0.1.0"
description = "Training-free text-to-text retrieval for remote sensing image retrieval (export)."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy>=1.23",
  "tqdm>=4.66",
  "pydantic>=2.6",
  "pyyaml>=6.0",
  "pillow>=10.0",
  "openai>=1.30.0",
]

[project.optional-dependencies]
local-vlm = ["torch>=2.1", "transformers>=4.40"]
local-embed = ["torch>=2.1", "transformers>=4.40", "sentence-transformers>=3.0"]

[project.scripts]
trsllava = "trsllava.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

